requirements.txt
requests
functools
flask
boto3
python-dotenv